requirements.txt
azure-common==1.1.4
azure-nspkg==1.0.0
azure-servicebus==0.20.3
requests==2.11.1
wsgiref==0.1.2
